Adding Find User and login screens classes

+++ Course 11 - Bank System/clsDeleteClientScreen.h
@@ -55,6 +55,11 @@ public:
 
 	static void ShowDeleteClientScreen()
 	{
+        if (!CheckAccessRights(clsUser::enPermissions::pDeleteClient))
+        {
+            return;// this will exit the function and it will not continue 
+        }
+
         clsScreen::_DrawScreenHeader("\tDelete Client Screen");
 
         DeleteClient();
